CopyProperties added, IncludeAll added, Bug fix

+++ BillingSoftwareBackend/BillingSoftware.Web/Controllers/AddressesController.cs
@@ -1,7 +1,9 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using BillingSoftware.Core.Contracts;
 using BillingSoftware.Core.Entities;
+using CommonBase.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -9,6 +11,7 @@ namespace BillingSoftware.Web.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class AddressesController : ControllerBase
     {
         private readonly IUnitOfWork _uow;
@@ -23,7 +26,11 @@ namespace BillingSoftware.Web.Controllers
         {
             try
             {
-                return Ok(await _uow.AddressRepository.GetAllAsync());
+                var email = HttpContext.User.Identity.Name;
+                var user = await _uow.UserRepository.GetUserByEmail(email);
+                var addresses = await _uow.AddressRepository.GetAllAsync();
+                addresses = addresses.Where(i => user.Company.Addresses.Any(a => a.Id == i.Id)).ToArray();
+                return Ok(addresses);
             }
             catch (System.Exception ex)
             {
@@ -31,27 +38,31 @@ namespace BillingSoftware.Web.Controllers
             }
         }
 
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Address>> GetAddress(int id)
-        {
-            try
-            {
-                var address = await _uow.AddressRepository.GetByIdAsync(id);
-                return address;
-            }
-            catch (System.Exception ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
+        //[HttpGet("{id}")]
+        //public async Task<ActionResult<Address>> GetAddress(int id)
+        //{
+        //    try
+        //    {
+        //        var address = await _uow.AddressRepository.GetByIdAsync(id);
+        //        return address;
+        //    }
+        //    catch (System.Exception ex)
+        //    {
+        //        return BadRequest(ex.Message);
+        //    }
+        //}
 
         [HttpPut]
         public async Task<IActionResult> PutAddress(Address address)
         {
             try
             {
+                if (!await CheckAuthorization(address.Id))
+                {
+                    return Unauthorized(new { Status = "Error", Message = $"You are not allowed to edit this address!" });
+                }
                 var entity = await _uow.AddressRepository.GetByIdAsync(address.Id);
-                entity.CopyProperties(address);
+                address.CopyProperties(entity);
                 _uow.AddressRepository.Update(entity);
                 await _uow.SaveChangesAsync();
                 return Ok();
@@ -67,6 +78,10 @@ namespace BillingSoftware.Web.Controllers
         {
             try
             {
+                if (!await CheckAuthorization(address.Id))
+                {
+                    return Unauthorized(new { Status = "Error", Message = $"You are not allowed to add this address!" });
+                }
                 await _uow.AddressRepository.AddAsync(address);
                 await _uow.SaveChangesAsync();
                 return Ok();
@@ -82,6 +97,11 @@ namespace BillingSoftware.Web.Controllers
         {
             try
             {
+                if (!await CheckAuthorization(id))
+                {
+                    return Unauthorized(new { Status = "Error", Message = $"You are not allowed to delete this address!" });
+                }
+
                 await _uow.AddressRepository.Remove(id);
                 await _uow.SaveChangesAsync();
                 return Ok();
@@ -90,6 +110,13 @@ namespace BillingSoftware.Web.Controllers
             {
                 return BadRequest(ex.Message);
             }
+        }
+
+        private async Task<bool> CheckAuthorization(int addressId)
+        {
+            var email = HttpContext.User.Identity.Name;
+            var user = await _uow.UserRepository.GetUserByEmail(email);
+            return user.Company.Addresses.Any(i => i.Id == addressId);
         }
     }
 }
