+++ BillingSoftwareBackend/BillingSoftware.Web/Controllers/ContactsController.cs
@@ -4,6 +4,8 @@ using System.Threading.Tasks;
 using BillingSoftware.Core.Contracts;
 using BillingSoftware.Core.Entities;
 using BillingSoftware.Persistence;
+using CommonBase.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -11,6 +13,7 @@ namespace BillingSoftware.Web.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class ContactsController : ControllerBase
     {
         private readonly IUnitOfWork _uow;
@@ -25,7 +28,11 @@ namespace BillingSoftware.Web.Controllers
         {
             try
             {
-                return Ok(await _uow.ContactRepository.GetAllAsync());
+                var email = HttpContext.User.Identity.Name;
+                var user = await _uow.UserRepository.GetUserByEmail(email);
+                var contacts = await _uow.ContactRepository.GetAllAsync();
+                contacts = contacts.Where(i => user.Company.Contacts.Any(a => a.Id == i.Id)).ToArray();
+                return Ok(contacts);
             }
             catch (System.Exception ex)
             {
@@ -33,27 +40,32 @@ namespace BillingSoftware.Web.Controllers
             }
         }
 
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Contact>> GetContact(int id)
-        {
-            try
-            {
-                var contact = await _uow.ContactRepository.GetByIdAsync(id);
-                return contact;
-            }
-            catch (System.Exception ex)
-            {
-                return BadRequest(ex.Message);
-            }
-        }
+        //[HttpGet("{id}")]
+        //public async Task<ActionResult<Contact>> GetContact(int id)
+        //{
+        //    try
+        //    {
+        //        var contact = await _uow.ContactRepository.GetByIdAsync(id);
+        //        return contact;
+        //    }
+        //    catch (System.Exception ex)
+        //    {
+        //        return BadRequest(ex.Message);
+        //    }
+        //}
 
         [HttpPut]
         public async Task<IActionResult> PutContact(Contact contact)
         {
             try
             {
+                if (!await CheckAuthorization(contact.Id))
+                {
+                    return Unauthorized(new { Status = "Error", Message = $"You are not allowed to edit this contact!" });
+                }
+
                 var entity = await _uow.ContactRepository.GetByIdAsync(contact.Id);
-                entity.CopyProperties(contact);
+                contact.CopyProperties(entity);
                 _uow.ContactRepository.Update(entity);
                 await _uow.SaveChangesAsync();
                 return Ok();
@@ -69,6 +81,11 @@ namespace BillingSoftware.Web.Controllers
         {
             try
             {
+                if (!await CheckAuthorization(contact.Id))
+                {
+                    return Unauthorized(new { Status = "Error", Message = $"You are not allowed to add a contact!" });
+                }
+
                 await _uow.ContactRepository.AddAsync(contact);
                 await _uow.SaveChangesAsync();
                 return Ok();
@@ -84,6 +101,11 @@ namespace BillingSoftware.Web.Controllers
         {
             try
             {
+                if (!await CheckAuthorization(id))
+                {
+                    return Unauthorized(new { Status = "Error", Message = $"You are not allowed to delete this contact!" });
+                }
+
                 await _uow.ContactRepository.Remove(id);
                 await _uow.SaveChangesAsync();
                 return Ok();
@@ -92,6 +114,13 @@ namespace BillingSoftware.Web.Controllers
             {
                 return BadRequest(ex.Message);
             }
+        }
+
+        private async Task<bool> CheckAuthorization(int contactId)
+        {
+            var email = HttpContext.User.Identity.Name;
+            var user = await _uow.UserRepository.GetUserByEmail(email);
+            return user.Company.Contacts.Any(i => i.Id == contactId);
         }
     }
 }
